Fix/parameters 20250425 (#4)
- version 3.12.0

- updated required Nextflow version to 24+

- simplified handling of input data
- revised and simplified parameters 
- limited published output to gene annotations, recombinase predictions, and mge predictions and sequences
- implemented various data flow control measures (success filtering etc.)

- recombinase_scan can now produce output in gff3 format
- recognise_genome now parses the specI information from the recognise results

- added CloWM-relevant files and documentation

- added experimental plasmid/contig mode (not functional)

+++ promgeflow/modules/get_db_seqs.nf
@@ -1,6 +1,7 @@
 process get_db_seqs {
     tag "${speci}"
     executor "local"
+    label "tiny"
     cpus 1
 
     input:
@@ -8,12 +9,16 @@ process get_db_seqs {
     val(seqdb_tar)
 
     output:
-    tuple val(speci), path("${speci}.ffn.gz"), emit: sequences
+    tuple val(speci), path("${speci}.ffn.gz"), emit: sequences, optional: true
+    tuple val(speci), path("${speci}.GETDBS_DONE"), emit: done_sentinel
 
     script:
     """
-    tar xvf ${seqdb_tar} ${speci}.genes.ffn.gz
-    mv -v ${speci}.genes.ffn.gz ${speci}.ffn.gz
+    tar xvf ${seqdb_tar} ${speci}.genes.ffn.gz || :
+    if [[ -f ${speci}.genes.ffn.gz ]]; then
+        mv -v ${speci}.genes.ffn.gz ${speci}.ffn.gz
+    fi
+    touch ${speci}.GETDBS_DONE
     """
     
 }
